feat: Enhance summarization capabilities and improve UI for float buttons; increase chunk retrieval and update styles

+++ modules/aiService.js
@@ -51,7 +51,7 @@ class AIService {
       const summarizer = await self.Summarizer.create({
         type: 'key-points',
         format: 'markdown',
-        length: 'short',
+        length: 'long',
         monitor: this.createMonitor(onProgress)
       });
 
@@ -77,7 +77,7 @@ class AIService {
       const summarizer = await self.Summarizer.create({
         type: 'key-points',
         format: 'markdown',
-        length: 'short'
+        length: 'long'
       });
 
       const stream = summarizer.summarizeStreaming(text, signal ? { signal } : {});
